move entity to entity folder and update readme with example
